began refactoring the home page
