+++ frontend/src/app/data-viewer/FieldLabel.tsx
@@ -1,0 +1,25 @@
+import { FONT_MONO } from './theme';
+import { Text } from '@mantine/core';
+
+export function FieldLabel({
+  children,
+  small,
+}: {
+  children: React.ReactNode;
+  small?: boolean;
+}) {
+  return (
+    <Text
+      style={{
+        fontFamily: FONT_MONO,
+        fontSize: small ? 10 : 11,
+        letterSpacing: '0.08em',
+        textTransform: 'uppercase',
+        color: 'rgba(246,245,239,0.5)',
+        marginBottom: 4,
+      }}
+    >
+      {children}
+    </Text>
+  );
+}
