+++ frontend/src/app/data-viewer/theme.ts
@@ -1,0 +1,15 @@
+export const COLOR = {
+  spruce: '#1B3A2F',
+  spruceDeep: '#122820',
+  slate: '#40525A',
+  birch: '#F6F5EF',
+  birchDim: '#EEEBE0',
+  ink: '#1B211D',
+  amber: '#dd9a2f',
+  amberSoft: '#E7B563',
+  amberYellow: '#FFD100',
+  line: 'rgba(27, 58, 47, 0.14)',
+};
+export const FONT_DISPLAY = "'Fraunces', 'Iowan Old Style', serif";
+export const FONT_BODY = "'General Sans', 'Inter', sans-serif";
+export const FONT_MONO = "'IBM Plex Mono', ui-monospace, monospace";
